+++ crates/cli/tests/reference/intrinsics.d.ts
@@ -1,0 +1,42 @@
+/* tslint:disable */
+/* eslint-disable */
+export function module_(): any;
+export function exports(): any;
+export function memory_(): any;
+export function is_memory_shared(): boolean;
+export function function_table(): any;
+
+export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembly.Module;
+
+export interface InitOutput {
+  readonly memory: WebAssembly.Memory;
+  readonly module_: () => any;
+  readonly exports: () => any;
+  readonly memory_: () => any;
+  readonly is_memory_shared: () => number;
+  readonly function_table: () => any;
+  readonly __wbindgen_export_0: WebAssembly.Table;
+  readonly __wbindgen_export_1: WebAssembly.Table;
+  readonly __wbindgen_start: () => void;
+}
+
+export type SyncInitInput = BufferSource | WebAssembly.Module;
+/**
+* Instantiates the given `module`, which can either be bytes or
+* a precompiled `WebAssembly.Module`.
+*
+* @param {{ module: SyncInitInput }} module - Passing `SyncInitInput` directly is deprecated.
+*
+* @returns {InitOutput}
+*/
+export function initSync(module: { module: SyncInitInput } | SyncInitInput): InitOutput;
+
+/**
+* If `module_or_path` is {RequestInfo} or {URL}, makes a request and
+* for everything else, calls `WebAssembly.instantiate` directly.
+*
+* @param {{ module_or_path: InitInput | Promise<InitInput> }} module_or_path - Passing `InitInput` directly is deprecated.
+*
+* @returns {Promise<InitOutput>}
+*/
+export default function __wbg_init (module_or_path?: { module_or_path: InitInput | Promise<InitInput> } | InitInput | Promise<InitInput>): Promise<InitOutput>;